[bugfix] Avoid vanishing predicates
Previously predicates, that could be hoisted up to a CallKernel
schedule item were silently dropped by the DummyHostASTBuilder.
This commit fixes the issue, by introducing a property whether
an AST builder is able to implement conditionals. If not so, the
hoisting of predicates is turned off (aka delayed until device
code is generated).