codegen: Ensure lower and upper slabs are distinct
When specifying split_iname with both lower and upper slabs, we need
to ensure that for loops shorter than the split length only one of the
two peeled slabs is run. Do so by ensuring that the upper slab is
conditioned on being larger than the end of the lower slab.

Fixes #215.