privatization for loops with non-zero base index (#429)
* privatization for loops with non-zero base index

Privatization was supported only for loops that had zero as their lower
bound. This commit fixes that by subtracting an offset of the iname's
lower bound expression. An option could have been to consider the new
shape of the temporary to be a function of iname's upper bound rather
than purely size, but that might need to unnecessary
register/stack-space usage.

* test privatization with non-zero lbound

Co-authored-by: Andreas Klöckner <inform@tiker.net>